+++ docs/changelog.rst
@@ -207,7 +207,7 @@ Artefact and Unscored epochs are now excluded from the calculation of the total 
 
 **New FAQ**
 
-The online documentation now has a brand new FAQ section! Make sure to check it out at https://raphaelvallat.com/yasa/build/html/faq.html
+The online documentation now has a brand new FAQ section! Make sure to check it out at https://raphaelvallat.com/yasa/faq.html
 
 **New function: coincidence matrix**